logout component Restructure
