+++ src/app/app.module.ts
@@ -77,7 +77,7 @@ import {TabViewModule} from 'primeng/tabview';
 import { LogoLoadingComponent } from './shared/logo-loading/logo-loading.component';
 import { TasksCardService } from './service/task/taskcard.service';
 import { ProjectEffects } from './main/dashboard/state/project.effects';
-import { LogoutComponent } from './main/logout/logout.component';
+import { LogoutComponent } from './main/dashboard/logout/logout.component';
 
 
 @NgModule({
